+++ src/utils/contentPreviewUpdate.js
