+++ src/utils/placeholderImage.js
