+++ src/utils/removeEmptyLines.js
